graph1

+++ MTrade/src/com/vkassin/mtrade/DayChartElement.java
@@ -6,7 +6,7 @@ import org.json.JSONObject;
 public class DayChartElement implements Comparable<DayChartElement> {
 
 	public Double open;
-	public Integer dateTime;
+	public Long dateTime;
 	public Integer volume;
 	public Double high;
 	public Double low;
@@ -16,7 +16,7 @@ public class DayChartElement implements Comparable<DayChartElement> {
 	public DayChartElement(JSONObject obj){
 
 		try{ this.open = obj.getDouble("open"); }catch(JSONException e){ }
-		try{ this.dateTime = obj.getInt("dateTime"); }catch(JSONException e){ }
+		try{ this.dateTime = obj.getLong("dateTime"); }catch(JSONException e){ }
 		try{ this.volume = obj.getInt("volume"); }catch(JSONException e){ }
 		try{ this.high = obj.getDouble("high"); }catch(JSONException e){ }
 		try{ this.low = obj.getDouble("low"); }catch(JSONException e){ }
